+++ Source/Scene/CentralBody.js
@@ -257,7 +257,7 @@ define([
 
     function computeDepthQuad(centralBody, frameState) {
         var radii = centralBody._ellipsoid.getRadii();
-        var p = frameState.camera.getPositionWC();
+        var p = frameState.camera.positionWC;
 
         // Find the corresponding position in the scaled space of the ellipsoid.
         var q = Cartesian3.multiplyComponents(centralBody._ellipsoid.getOneOverRadii(), p);
@@ -687,7 +687,7 @@ define([
             this._hasWaterMask = hasWaterMask;
         }
 
-        var cameraPosition = frameState.camera.getPositionWC();
+        var cameraPosition = frameState.camera.positionWC;
 
         this._occluder.setCameraPosition(cameraPosition);