+++ Source/Scene/SkyAtmosphere.js
@@ -171,7 +171,7 @@ define([
             this._spSkyFromAtmosphere = shaderCache.getShaderProgram(vs, SkyAtmosphereFS);
         }
 
-        var cameraPosition = frameState.camera.getPositionWC();
+        var cameraPosition = frameState.camera.positionWC;
 
         this._fCameraHeight2 = Cartesian3.magnitudeSquared(cameraPosition);
         this._fCameraHeight = Math.sqrt(this._fCameraHeight2);